[refactor/#53] CreateTodo 모달 수정
